Addressed more comments from the pr
