+++ pointmatcher/ICP.cpp
@@ -43,7 +43,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "MatchersImpl.h"
 #include "OutlierFiltersImpl.h"
 #include "ErrorMinimizersImpl.h"
-#include "PointToPlaneErrorMinimizer.h"
+#include "ErrorMinimizers/PointToPlane.h"
 #include "TransformationCheckersImpl.h"
 #include "InspectorsImpl.h"
